Clean up code [CLEANUP], document ben_list_for_each() [DOCUMENTATION]

+++ include/bencodetools/bencode.h
@@ -575,9 +575,18 @@ static inline const char *ben_str_val(const struct bencode *b)
  * using ben_list_pop_current().
  *
  * pos is a size_t.
+ *
+ * Example:
+ *
+ * size_t pos;
+ * struct bencode *list = xxx;
+ * struct bencode *value;
+ * ben_list_for_each(value, pos, list) {
+ *         inspect(value);
+ * }
  */
 #define ben_list_for_each(value, pos, l) \
-	for ((pos) = 0; \
+	for ((pos) = (size_t) 0;		    \
 	     (pos) < (ben_list_const_cast(l))->n && \
 	     ((value) = ((const struct bencode_list *) (l))->values[(pos)]) != NULL ; \
 	     (pos)++)
@@ -591,11 +600,12 @@ static inline const char *ben_str_val(const struct bencode *b)
  * Filter out all items from list whose string value does not begin with "foo".
  *
  * ben_list_for_each(value, pos, list) {
- *     if (strncmp(ben_str_val(value), "foo", 3) != 0)
- *         ben_free(ben_list_pop_current(&pos, list));
+ *         if (strncmp(ben_str_val(value), "foo", 3) != 0)
+ *                 ben_free(ben_list_pop_current(&pos, list));
  * }
  */
-static inline struct bencode *ben_list_pop_current(struct bencode *list, size_t *pos)
+static inline struct bencode *ben_list_pop_current(struct bencode *list,
+						   size_t *pos)
 {
 	struct bencode *value = ben_list_pop(list, *pos);
 	(*pos)--;
